Добавил систему уведомлений, проект запускается к 1с можно подключится,  но система не проверенна на работоспособность

+++ haspd/entrypoint.sh
@@ -1,21 +1,27 @@
 #!/bin/bash
 set -e
 
+: "${NOTIFY_SH:?❌ NOTIFY_SH не задан!}" 
+SCRIPT_NAME="entrypoint.sh (haspd)"
+source ${NOTIFY_SH}
+trap 'handle_exit' EXIT
+
 echo "🧹 Очищаем логи и временные файлы..."
 find "$LOG_DIR" -type f -name "*.log" -exec truncate -s 0 {} \;
 rm -rf /tmp/.X* /tmp/.X11-unix /root/.vnc/*.pid
 
-: "${LOG_DIR:?❌ LOG_DIR не задан! Проверь переменные окружения.}"
+: "${LOG_DIR:?❌ LOG_DIR не задан! Проверь переменные окружения.}" 
  
 # Проверка и загрузка VNC-пароля
-if [ ! -s "$VNC_PASSWORD_FILE" ]; then
-  echo "❌ Файл VNC_PASSWORD_FILE пуст или не существует — проверь маунт секрета" >&2
+if [ ! -s "$VNC_PASSWORD_FILE" ]; then 
+  LAST_ERROR_MESSAGE="❌ Файл VNC_PASSWORD_FILE пуст или не существует — проверь маунт секрета"
+  echo "$LAST_ERROR_MESSAGE" >&2
   exit 1
 else
   export VNC_PASSWORD=$(cat "$VNC_PASSWORD_FILE")
   echo "🔐 Пароль от VNC успешно загружен из секрета"
 fi
-
+ 
 # Создание файла xstartup
 echo "⚙️ Настройка xstartup..."
 cat <<EOF > /root/.vnc/xstartup
